Show tooltip when hovering over sample name/description, not just label
